backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.30
pydantic==2.7.0
pydantic-settings==2.2.1
mysqlclient==2.2.4
bcrypt==4.1.3
PyJWT==2.8.0
python-multipart==0.0.9